% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmuer.R
\name{nmuer_weights_constant}
\alias{nmuer_weights_constant}
\alias{nmuer_weights_logistic}
\title{Weighting functions for unified estimate}
\usage{
nmuer_weights_constant(observation_vector = NULL, DATA = NULL)

nmuer_weights_logistic(missing_formula, observation_vector = NULL, DATA = NULL)
}
\arguments{
\item{observation_vector}{A vector of TRUE and FALSE indicating
if the ith row has all variables in the analysis or working model
observed. This is automatically supplied by nmuer().}

\item{DATA}{A dataframe on which the weighting function will
be fit to calculate the weights. This is automatically supplied
by nmuer()}

\item{missing_formula}{A formula object, the right-hand side defines
the variables to be used in the logistic model}
}
\description{
Functions to supply inverse probability weights for unified
estimates
}
\details{
When missingness is MAR, using inverse probability weights (ipw) can
be used to rebalance the estimating equations to allow for unbiased
estimation of parameters. These functions are designed to allow 
nmuer() to automatically estimate the weights. Users can also write
their own weight functions.

Each weight function MUST take an argument called DATA and an argument
called observation_vector. The observation_vector and DATA should be 
left NULL, the nmuer() code automatically calculates and includes them.
Other arguments may be defined depending on the method of estimating 
weights.

Each weight function must output a vector of weights the same
length as the number of rows in DATA. These weights are then
supplied to the model fitting functions glm() and coxph().

nmuer_weights_constant() gives all observations the same weight.
This is appropriate for MCAR data. The user does not need to
enter any additional information to use this method.

nmuer_weights_logistic() uses a GLM model with a binomial family and
a logistic link function to estimate the probability a row has complete
data, then takes the inverse of that probability as the weight. A
missing_formula argument is required. The right-hand side of the 
formula will be used to model observation probability. The left-hand
side of the formula will be automatically replaced by nmuer() to use
the appropriate observation vector. If the working_model has no
missing data, then nmuer_weights_logistic() will return a vector of 1s
with a warning.
}
